backend/requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# LangChain & LLM - Using compatible versions
langchain==0.3.12
langchain-groq==0.2.5  # Latest stable 0.2.x version
langchain-core>=0.3.25  # Required by langchain 0.3.12
langsmith==0.2.4

# Environment & Utils
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Text-to-Speech
gTTS==2.5.3

# Core LangGraph
langgraph>=0.2.0
langgraph-checkpoint>=1.0.0